Break down init_deck into smaller modules
